tested tree shakability at a react project
