+++ examples/react/develop/simple-counter/package.json
@@ -3,8 +3,8 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
-    "@agile-ts/core": "^0.0.17",
-    "@agile-ts/react": "^0.1.0",
+    "@agile-ts/core": "file:.yalc/@agile-ts/core",
+    "@agile-ts/react": "file:.yalc/@agile-ts/react",
     "react": "17.0.2",
     "react-dom": "17.0.2",
     "react-scripts": "4.0.0"
@@ -17,9 +17,9 @@
     "build": "react-scripts build",
     "test": "react-scripts test",
     "eject": "react-scripts eject",
-    "analyze": "source-map-explorer 'build/static/js/*.js'",
-    "install:dev:agile": "yalc add @agile-ts/core @agile-ts/react @agile-ts/logger & yarn install",
-    "install:prod:agile": "yarn add @agile-ts/core @agile-ts/react @agile-ts/logger & yarn install"
+    "analyze": "yarn run build && source-map-explorer 'build/static/js/*.js'",
+    "install:dev:agile": "yalc add @agile-ts/core @agile-ts/react & yarn install",
+    "install:prod:agile": "yarn add @agile-ts/core @agile-ts/react & yarn install"
   },
   "browserslist": {
     "production": [
